Added test for ParserDoubleItem

+++ opm/parser/eclipse/Parser/ParserDoubleItem.cpp
@@ -34,7 +34,7 @@ namespace Opm {
     }
 
 
-    ParserDoubleItem::ParserDoubleItem(const std::string& itemName, ParserItemSizeEnum sizeType, int defaultValue) : ParserItem(itemName, sizeType) {
+    ParserDoubleItem::ParserDoubleItem(const std::string& itemName, ParserItemSizeEnum sizeType, double defaultValue) : ParserItem(itemName, sizeType) {
         m_default = defaultValue;
     }
